separate UPnP from Game/Server into it's own class
