+++ OpenRA.Game/OpenRA.Game.csproj
@@ -222,6 +222,7 @@
     <Compile Include="Network\ReplayRecorderConnection.cs" />
     <Compile Include="Widgets\TooltipContainerWidget.cs" />
     <Compile Include="Traits\DebugPauseState.cs" />
+    <Compile Include="Network\UPnP.cs" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\OpenRA.FileFormats\OpenRA.FileFormats.csproj">
